solr changes and rebuild

+++ client/openstorefront/app/scripts/services/componentservice.js
@@ -70,7 +70,7 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
     var result = $q.defer();
     if (id && reviewId && pro)
     {
-      var url = 'api/v1/resource/components/'+id+'/review/'+reviewId+'/pro';
+      var url = 'api/v1/resource/components/'+id+'/reviews/'+reviewId+'/pro';
       $http({
         method: 'DELETE',
         url: url,
@@ -78,7 +78,7 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
       })
       .success(function(data, status, headers, config) { /*jshint unused:false*/
       });
-      url = 'api/v1/resource/components/'+id+'/review/'+reviewId+'/con';
+      url = 'api/v1/resource/components/'+id+'/reviews/'+reviewId+'/con';
       $http({
         method: 'DELETE',
         url: url,
@@ -106,7 +106,7 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
     var result = $q.defer();
     if (id && questionId)
     {
-      var url = 'api/v1/resource/components/'+id+'/question/'+questionId;
+      var url = 'api/v1/resource/components/'+id+'/questions/'+questionId;
       $http({
         method: 'DELETE',
         url: url,
@@ -137,10 +137,10 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
       var methodString; 
       if (questionId){
         methodString = 'PUT';
-        url = 'api/v1/resource/components/'+id+'/question/'+questionId;
+        url = 'api/v1/resource/components/'+id+'/questions/'+questionId;
       } else {
         methodString = 'POST';
-        url = 'api/v1/resource/components/'+id+'/question';
+        url = 'api/v1/resource/components/'+id+'/questions';
       }
       $http({
         method: methodString,
@@ -166,11 +166,11 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
   }
 
 
-  componentservice.deleteResponse = function(id, responseId) {
+  componentservice.deleteResponse = function(id, responseId, questionId) {
     var result = $q.defer();
-    if (id && responseId)
+    if (id && responseId && questionId)
     {
-      var url = 'api/v1/resource/components/'+id+'/response/'+responseId;
+      var url = 'api/v1/resource/components/'+id+'/questions/'+questionId+'/responses/'+responseId;
       $http({
         method: 'DELETE',
         url: url,
@@ -200,10 +200,10 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
       var url;
       var methodString;
       if (!responseId) {
-        url = 'api/v1/resource/components/'+id+'/response/'+questionId;
+        url = 'api/v1/resource/components/'+id+'/questions/'+questionId+'/responses';
         methodString = 'POST';
       } else {
-        url = 'api/v1/resource/components/'+id+'/response/'+responseId;
+        url = 'api/v1/resource/components/'+id+'/questions/'+questionId+'/responses/'+responseId;
         methodString = 'PUT';
       }
       $http({
@@ -233,7 +233,7 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
     var result = $q.defer();
     if (id && reviewId)
     {
-      var url = 'api/v1/resource/components/'+id+'/review/'+reviewId;
+      var url = 'api/v1/resource/components/'+id+'/reviews/'+reviewId;
       $http({
         method: 'DELETE',
         url: url,
@@ -267,10 +267,10 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
       var url;
       var methodString;
       if (!reviewId) {
-        url = 'api/v1/resource/components/'+id+'/review';
+        url = 'api/v1/resource/components/'+id+'/reviews';
         methodString = 'POST';
       } else {
-        url = 'api/v1/resource/components/'+id+'/review/' + reviewId;
+        url = 'api/v1/resource/components/'+id+'/reviews/' + reviewId;
         methodString = 'PUT';
       }
       $http({
@@ -300,7 +300,7 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
     var result = $q.defer();
     if (id && reviewId && pro)
     {
-      var url = 'api/v1/resource/components/'+id+'/review/'+reviewId+'/pro';
+      var url = 'api/v1/resource/components/'+id+'/reviews/'+reviewId+'/pros';
       $http({
         method: 'POST',
         url: url,
@@ -328,7 +328,7 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
     var result = $q.defer();
     if (id && reviewId && con)
     {
-      var url = 'api/v1/resource/components/'+id+'/review/'+reviewId + '/con';
+      var url = 'api/v1/resource/components/'+id+'/reviews/'+reviewId + '/cons';
       $http({
         method: 'POST',
         url: url,
@@ -371,7 +371,7 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
         .success(function(data, status, headers, config) { /*jshint unused:false*/          
           if (data && !isEmpty(data) && isNotRequestError(data)) {
             removeError();
-            console.log('data', data);
+            // console.log('data', data);
             save('component_'+id, data);
             result.resolve(data);
           } else {
@@ -437,8 +437,75 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
     return result.promise;
   };
 
-  componentservice.doSearch = function(type, key) {
-    return componentservice.getComponentList();
+  componentservice.doSearch = function(type, key, architecture) {
+    var result            = $q.defer();
+    var url               = 'api/v1/service/search/all';
+    var value             = null;
+    var cachedComponents  = null;
+    var name;
+
+    
+    if (type && key) {
+      type  = type.toLowerCase();
+      if (type !== 'attribute'){
+        key   = key.toLowerCase();
+        name  = type + '-' + key;
+      } else {
+        if (!key.key) {
+          name = type + key.type.toLowerCase() + '-ALL';
+        } else {
+          name = type + key.type.toLowerCase() + '-' + key.key.toLowerCase();
+        }
+      }
+      // if they don't give me an ID I send them back the whole list.
+      value = checkExpire(name, minute * 10);
+      cachedComponents = checkExpire('cachedComponents', minute * 1440); // a day
+      // console.log('started search');
+      // var start = new Date().getTime();
+      if (value) {
+        // var end = new Date().getTime();
+        // var time = end - start;
+        // console.log('Total Execution time: ' + time);
+        result.resolve(value);
+      } else {
+        if (cachedComponents) {
+          var temp = search({'type': type, 'key': key}, cachedComponents, architecture);
+          // var end = new Date().getTime();
+          // var time = end - start;
+          // console.log('finished search!', temp);
+          // console.log('Total Execution time: ' + time);
+          save(name, temp);
+          result.resolve(temp);
+        } else {
+          $http({
+            method: 'GET',
+            url: url
+          })
+          .success(function(data, status, headers, config) { /*jshint unused:false*/
+            if (data && !isEmpty(data) && isNotRequestError(data)) {
+              removeError();
+              save('cachedComponents', data);
+              var temp = search({'type': type, 'key': key}, data, architecture);
+              // var end = new Date().getTime();
+              // var time = end - start;
+              // console.log('finished search!', temp);
+              // console.log('Total Execution time: ' + time);
+              save(name, temp);
+              result.resolve(temp);
+            } else {
+              removeError();
+              triggerError(data);
+              result.reject(false);
+            }
+          }).error(function(data, status, headers, config){
+            result.reject('There was a server error');
+          });
+        }
+      }
+    } else {
+      result.reject('You must provide a type and key for the search');
+    }
+    return result.promise;
   };
 
 
@@ -472,7 +539,7 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
     $http.delete('api/v1/resource/components/'+id+'/tags');
     $http({
       method: 'POST',
-      url: 'api/v1/resource/components/'+id+'/tags',
+      url: 'api/v1/resource/components/'+id+'/tags/list',
       data: tags
     }).success(function(data, status, headers, config){
       if (data && data !== 'false' && isNotRequestError(data)) {
@@ -489,5 +556,57 @@ app.factory('componentservice', ['$http', '$q', 'localCache', function($http, $q
     return deferred.promise;
   };
 
+  componentservice.addTag = function(id, tag) {
+    var deferred = $q.defer();
+    if (id && tag) {
+      $http({
+        method: 'POST',
+        url: 'api/v1/resource/components/'+id+'/tags',
+        data: tag
+      }).success(function(data, status, headers, config){
+        if (data && data !== 'false' && isNotRequestError(data)) {
+          removeError();
+          deferred.resolve(data);
+        } else {
+          removeError();
+          triggerError(data);
+          deferred.reject(false);
+        }
+      }).error(function(data, status, headers, config){
+        deferred.resolve('There was an error saving the tags');
+      });
+    } else {
+      deferred.reject('You need a component Id and a tag to save.')
+    }
+    return deferred.promise;
+  };
+
+  componentservice.removeTag = function(id, tag) {
+    var deferred = $q.defer();
+    if (id && tag) {
+      $http({
+        method: 'DELETE',
+        url: 'api/v1/resource/components/'+id+'/tags/text',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        data: tag
+      }).success(function(data, status, headers, config){
+        if (data !== 'false' && isNotRequestError(data)) {
+          removeError();
+          deferred.resolve(data);
+        } else {
+          removeError();
+          triggerError(data);
+          deferred.reject(false);
+        }
+      }).error(function(data, status, headers, config){
+        deferred.reject(false);
+      });
+    } else {
+      deferred.reject(false);
+    }
+    return deferred.promise;
+  }
   return componentservice;
 }]);
