+++ server/openstorefront/openstorefront-web/src/main/java/edu/usu/sdl/openstorefront/web/action/ResourceAction.java
@@ -27,6 +27,7 @@ import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.Map;
 import javax.servlet.http.HttpServletResponse;
+import net.sourceforge.stripes.action.ErrorResolution;
 import net.sourceforge.stripes.action.FileBean;
 import net.sourceforge.stripes.action.HandlesEvent;
 import net.sourceforge.stripes.action.Resolution;
@@ -69,7 +70,7 @@ public class ResourceAction
 			protected void stream(HttpServletResponse response) throws Exception
 			{
 				Path path = componentResource.pathToResource();
-				if (path != null) {
+				if (path != null && path.toFile().exists()) {
 					Files.copy(path, response.getOutputStream());
 				} else {
 					throw new OpenStorefrontRuntimeException("Resource not on disk", "Check resource record: " + resourceId);
@@ -83,27 +84,30 @@ public class ResourceAction
 	public Resolution uploadResource()
 	{
 		Map<String, String> errors = new HashMap<>();
-		if (componentResource != null) {
-			componentResource.setActiveStatus(ComponentResource.ACTIVE_STATUS);
-			componentResource.setUpdateUser(SecurityUtil.getCurrentUserName());
-			componentResource.setCreateUser(SecurityUtil.getCurrentUserName());
-			componentResource.setOriginalName(file.getFileName());
-			componentResource.setMimeType(file.getContentType());
+		if (SecurityUtil.isAdminUser()) {
+			if (componentResource != null) {
+				componentResource.setActiveStatus(ComponentResource.ACTIVE_STATUS);
+				componentResource.setUpdateUser(SecurityUtil.getCurrentUserName());
+				componentResource.setCreateUser(SecurityUtil.getCurrentUserName());
+				componentResource.setOriginalName(file.getFileName());
+				componentResource.setMimeType(file.getContentType());
 
-			ValidationModel validationModel = new ValidationModel(componentResource);
-			validationModel.setConsumeFieldsOnly(true);
-			ValidationResult validationResult = ValidationUtil.validate(validationModel);
-			if (validationResult.valid()) {
-				try {
-					service.getComponentService().saveResourceFile(componentResource, file.getInputStream());
-				} catch (IOException ex) {
-					throw new OpenStorefrontRuntimeException("Unable to able to save resource.", "Contact System Admin. Check disk space and permissions.", ex);
+				ValidationModel validationModel = new ValidationModel(componentResource);
+				validationModel.setConsumeFieldsOnly(true);
+				ValidationResult validationResult = ValidationUtil.validate(validationModel);
+				if (validationResult.valid()) {
+					try {
+						service.getComponentService().saveResourceFile(componentResource, file.getInputStream());
+					} catch (IOException ex) {
+						throw new OpenStorefrontRuntimeException("Unable to able to save resource.", "Contact System Admin. Check disk space and permissions.", ex);
+					}
+				} else {
+					errors.put("file", validationResult.toHtmlString());
 				}
-			} else {
-				errors.put("file", validationResult.toHtmlString());
 			}
+			return streamUploadResponse(errors);
 		}
-		return streamUploadResponse(errors);
+		return new ErrorResolution(HttpServletResponse.SC_FORBIDDEN, "Access denyed");
 	}
 
 	public String getResourceId()
